Complete checkpoint w/ abstracted code & working tests
