+++ spec/apprentice_teacher_spec.rb
@@ -1,4 +1,5 @@
 require_relative '../apprentice_teacher'
+require_relative '../give_props'
 
 describe ApprenticeTeacher do
   let(:rey) { ApprenticeTeacher.new({name: "Rey", age: 21}) }
