+++ student.rb
@@ -1,16 +1,12 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'give_props'
+
+class Student < GiveProps
 
   def initialize(options = {})
     @phase = 1
-    @age = options.fetch(:age, 0)
     @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
 
   def set_phase(num)
     response = ""
@@ -21,7 +17,7 @@ class Student
       response = "Oooh, phase #{num}. I hope I'm ready!"
     end
     @phase = num
-    response
+    response  
   end
 
   def learn_stuff
